Updated message

+++ backend/Controllers/comment.js
@@ -18,7 +18,7 @@ exports.addComment = async (req, res) => {
   } catch (err) {
     res.status(500).json({
       success: false,
-      error: err.message,
+      msg: err.message,
     });
   }
 };
@@ -38,7 +38,7 @@ exports.getCommentByVideoId = async (req, res) => {
   } catch (err) {
     res.status(500).json({
       success: false,
-      error: err.message,
+      msg: err.message,
     });
   }
 };
@@ -54,9 +54,17 @@ exports.editComment = async (req, res) => {
     );
     if (!comment) {
       return res.status(404).json({
-        error: "Comment not found",
+        msg: "Comment not found",
       });
     }
+
+     // Check if the logged-in user is the owner of the comment
+     if (comment.user._id.toString() !== req.user._id.toString()) {
+      return res.status(403).json({
+        msg: "You are not authorized to edit this comment",
+      });
+    }
+
     comment.message = message;  // updates the comment message
     await comment.save();
     res.status(200).json({
@@ -65,7 +73,7 @@ exports.editComment = async (req, res) => {
     });
   } catch (err) {
     res.status(500).json({
-      error: "Internal server error",
+      msg: "Internal server error",
     });
   }
 };
@@ -80,16 +88,25 @@ exports.deleteComment = async (req, res) => {
     );
     if (!comment) {
       return res.status(404).json({
-        error: "Comment not found",
+        msg: "Comment not found",
       });
     }
+
+    // Check if the logged-in user is the owner of the comment
+    if (comment.user._id.toString() !== req.user._id.toString()) {
+      return res.status(403).json({
+        msg: "You are not authorized to delete this comment",
+      });
+    }
+
+
     await comment.deleteOne({ _id: videoId });  // delete the comment from database
     res.status(200).json({
       msg: "Comment deleted",
     });
   } catch (err) {
     res.status(500).json({
-      error: "Internal server error",
+      msg: "Internal server error",
     });
   }
 };
